requirements.txt
bleach==6.2.0
blinker==1.9.0
boto3==1.35.66
botocore==1.35.66
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cycler==0.12.1
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.0
fsspec==2024.10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kaggle==1.6.17
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.32.3
s3transfer==0.10.4
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.6.0
six==1.16.0
sympy==1.13.1
text-unidecode==1.3
threadpoolctl==3.5.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.1.3
